Move CLI commands to a "cli" module
