+++ setup.py
@@ -162,16 +162,16 @@ setup(
     },
     entry_points={
         'console_scripts': [
-            'girder-server = girder.__main__:main',
+            'girder-server = girder.cli.serve:main',
             'girder-install = girder.utility.install:main',
-            'girder-sftpd = girder.api.sftp:main',
-            'girder-shell = girder.utility.shell:main',
+            'girder-sftpd = girder.cli.sftpd:main',
+            'girder-shell = girder.cli.shell:main',
             'girder = girder.cli:main'
         ],
         'girder.cli_plugins': [
-            'serve = girder.__main__:main',
-            'shell = girder.utility.shell:main',
-            'sftpd = girder.api.sftp:main'
+            'serve = girder.cli.serve:main',
+            'shell = girder.cli.shell:main',
+            'sftpd = girder.cli.sftpd:main'
         ]
     }
 )
